error-reporter: extract SentryClient to a separate file for testability.
