+++ packages/buidler-core/src/internal/error-reporter/error-reporter.ts
@@ -1,6 +1,4 @@
-import * as Sentry from "@sentry/node";
 import debug from "debug";
-import os from "os";
 
 import { BuidlerError, BuidlerPluginError } from "../core/errors";
 import { REVERSE_ERRORS_MAP } from "../core/errors-list";
@@ -11,10 +9,11 @@ import {
   getUserAgent,
   getUserType,
   isLocalDev,
-  UserType,
 } from "../util/analytics";
 
-interface ErrorContextData {
+import { ErrorReporterClient, SentryClient } from "./sentry";
+
+export interface ErrorContextData {
   errorType: "BuidlerError" | "BuidlerPluginError" | "Error";
   // true if is originated from Buidler, false otherwise
   isBuidlerError: boolean;
@@ -50,14 +49,6 @@ interface ErrorContextData {
 }
 
 /**
- * The client that actually sends the error report
- */
-interface ErrorReporterClient {
-  sendMessage(message: string, context: any): Promise<void>;
-  sendErrorReport(error: Error): Promise<void>;
-}
-
-/**
  * The ErrorReporter interface that is exposed publicly.
  */
 interface ErrorReporterInterface {
@@ -65,114 +56,8 @@ interface ErrorReporterInterface {
   sendErrorReport(error: Error): Promise<void>;
 }
 
-class SentryClient implements ErrorReporterClient {
-  public static SENTRY_FLUSH_TIMEOUT = 3000;
-  private readonly _SENTRY_DSN =
-    "https://38ba58bb85fa409e9bb7f50d2c419bc2@o385026.ingest.sentry.io/5224869";
-  private readonly _log = debug("buidler:core:error-reporter:sentry");
-
-  constructor(
-    projectId: string,
-    clientId: string,
-    userType: UserType,
-    userAgent: string,
-    buidlerVersion: string
-  ) {
-    // init Sentry client
-    Sentry.init({ dsn: this._SENTRY_DSN });
-
-    // setup metadata to be included in all reports by default
-    Sentry.configureScope((scope) => {
-      scope.setUser({ id: clientId, type: userType });
-      scope.setTag("projectId", projectId);
-      scope.setTag("version", buidlerVersion);
-      scope.setTag("os", os.type());
-      scope.setTag("node", process.version);
-      scope.setTag("userAgent", userAgent);
-      scope.setExtra("platform", os.platform());
-      scope.setExtra("os release", os.release());
-    });
-
-    this._log("Sentry client init");
-  }
-
-  public async sendMessage(message: string, context: any) {
-    this._log("Sending message...", { message, context });
-
-    Sentry.withScope(function (scope) {
-      scope.setExtras(context);
-
-      Sentry.captureMessage(message);
-    });
-
-    try {
-      await Sentry.flush(SentryClient.SENTRY_FLUSH_TIMEOUT);
-      this._log("Message sent");
-    } catch (error) {
-      // absorb the sentry error and log it
-      this._log(`Could not send message. Reason: `, error.message || error);
-    }
-  }
-
-  public async sendErrorReport(error: Error): Promise<void> {
-    this._log("Sending error report...");
-    const errorContextData = contextualizeError(error);
-
-    const {
-      errorType,
-      pluginName,
-      title,
-      description,
-      name,
-      number,
-      message,
-      category,
-      contextMessage,
-    } = errorContextData;
-
-    Sentry.withScope(function (scope) {
-      scope.setTag("errorType", errorType);
-      scope.setExtra("message", message);
-      if (pluginName !== undefined) {
-        scope.setTag("pluginName", pluginName);
-      }
-      if (name !== undefined) {
-        scope.setTag("name", name);
-      }
-      if (number !== undefined) {
-        scope.setTag("number", String(number));
-      }
-      if (title !== undefined) {
-        scope.setExtra("title", title);
-      }
-      if (contextMessage !== undefined) {
-        scope.setExtra("contextMessage", contextMessage);
-      }
-      if (category !== undefined) {
-        scope.setTag("category.name", category.name);
-        scope.setExtra("category.title", category.title);
-      }
-      if (description !== undefined) {
-        scope.setExtra("description", description);
-      }
-
-      Sentry.captureException(error);
-    });
-
-    try {
-      await Sentry.flush(SentryClient.SENTRY_FLUSH_TIMEOUT);
-      this._log(`Successfully sent report: '${message}'`);
-    } catch (error) {
-      // absorb the sentry error and log it
-      this._log(
-        `Could not sent report for error '${message}', Reason: `,
-        error.message || error
-      );
-    }
-  }
-}
-
 const log = debug(`buidler:core:error-reporter`);
+
 export class ErrorReporter implements ErrorReporterInterface {
   /**
    * Setup ErrorReporter instance.
@@ -200,13 +85,15 @@ export class ErrorReporter implements ErrorReporterInterface {
     const userType = getUserType();
     const userAgent = getUserAgent();
 
-    this._instance = new ErrorReporter({
+    const client = new SentryClient(
       projectId,
       clientId,
       userType,
       userAgent,
-      buidlerVersion,
-    });
+      buidlerVersion
+    );
+
+    this._instance = new ErrorReporter(client);
   }
 
   /**
@@ -226,26 +113,8 @@ export class ErrorReporter implements ErrorReporterInterface {
 
   private readonly _client: ErrorReporterClient;
 
-  private constructor({
-    projectId,
-    clientId,
-    userType,
-    userAgent,
-    buidlerVersion,
-  }: {
-    projectId: string;
-    clientId: string;
-    userType: UserType;
-    userAgent: string;
-    buidlerVersion: string;
-  }) {
-    this._client = new SentryClient(
-      projectId,
-      clientId,
-      userType,
-      userAgent,
-      buidlerVersion
-    );
+  private constructor(client: ErrorReporterClient) {
+    this._client = client;
   }
 
   public async sendMessage(message: string, context: any) {
@@ -253,7 +122,8 @@ export class ErrorReporter implements ErrorReporterInterface {
   }
 
   public async sendErrorReport(error: Error) {
-    await this._client.sendErrorReport(error);
+    const errorContext = contextualizeError(error);
+    await this._client.sendErrorReport(error, errorContext);
   }
 
 }
